Move Xbox DSound types into its own header file.
